it actually works
git-svn-id: https://localhost/igvc@747 b8439097-6f30-0410-869d-b57f63bf1bb9

+++ 2010/software/trunk/newencoders/pcdriver/quadCoderDriver.cc
@@ -6,7 +6,7 @@
 //#define WHEEL_RADIUS		((double)5 / (double)MOTOR_RATIO)
 #define WHEEL_RADIUS		(double(.146) / double(2))
 #define WHEEL_BASE		((double)28)
-#define NUMBERTICKS		double(200)
+#define NUMBERTICKS		double(800)
 
 #define SMAPLEDELAY		double(5e-3)
 quadCoderDriver::quadCoderDriver()
